+++ packages/gi-assets-basic/src/components/EdgeMerge/registerMeta.ts
@@ -1,0 +1,14 @@
+import { extra } from '@antv/gi-sdk';
+import info from './info';
+const { deepClone, GIAC_METAS } = extra;
+
+const metas = deepClone(GIAC_METAS);
+metas.GIAC.properties.GIAC.properties.title.default = info.name;
+metas.GIAC.properties.GIAC.properties.icon.default = info.icon;
+metas.GIAC.properties.GIAC.properties.tooltip.default = info.desc;
+metas.GIAC.properties.GIAC.properties.isShowTooltip.default = true;
+
+const registerMeta = () => {
+  return metas;
+};
+export default registerMeta;