+++ packages/gi-assets-basic/src/components/RichContainer/registerMeta.ts
@@ -1,13 +1,14 @@
+import $i18n from '../../i18n';
 const registerMeta = context => {
   const { GIAC_CONTENT_ITEMS = [], GIAC_ITEMS = [] } = context;
   return {
     containers: [
       {
         id: 'navbar-left',
-        name: '导航左区',
+        name: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.NavigationLeftArea', dm: '导航左区' }),
         required: true,
         GI_CONTAINER: {
-          title: '集成组件',
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.IntegratedComponents', dm: '集成组件' }),
           type: 'string',
           'x-decorator': 'FormItem',
           'x-component': 'Select',
@@ -20,10 +21,10 @@ const registerMeta = context => {
       },
       {
         id: 'navbar-right',
-        name: '导航右区',
+        name: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.NavigationRightArea', dm: '导航右区' }),
         required: true,
         GI_CONTAINER: {
-          title: '集成组件',
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.IntegratedComponents', dm: '集成组件' }),
           type: 'string',
           'x-decorator': 'FormItem',
           'x-component': 'Select',
@@ -36,10 +37,10 @@ const registerMeta = context => {
       },
       {
         id: 'view-mode',
-        name: '模式展示',
+        name: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.ModeDisplay', dm: '模式展示' }),
         required: true,
         GI_CONTAINER: {
-          title: '集成组件',
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.IntegratedComponents', dm: '集成组件' }),
           type: 'string',
           'x-decorator': 'FormItem',
           'x-component': 'Select',
@@ -48,14 +49,21 @@ const registerMeta = context => {
           },
           enum: GIAC_CONTENT_ITEMS,
           default: [],
+        },
+        icon: {
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.DisplayIcon', dm: '展示图标' }),
+          type: 'string',
+          'x-decorator': 'FormItem',
+          'x-component': 'Input',
+          default: 'icon-deploymentunit1',
         },
       },
       {
         id: 'data-query',
-        name: '数据查询',
+        name: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.DataQuery', dm: '数据查询' }),
         required: true,
         GI_CONTAINER: {
-          title: '集成组件',
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.IntegratedComponents', dm: '集成组件' }),
           type: 'string',
           'x-decorator': 'FormItem',
           'x-component': 'Select',
@@ -65,13 +73,43 @@ const registerMeta = context => {
           enum: GIAC_CONTENT_ITEMS,
           default: [],
         },
+        icon: {
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.DisplayIcon', dm: '展示图标' }),
+          type: 'string',
+          'x-decorator': 'FormItem',
+          'x-component': 'Input',
+          default: 'icon-query',
+        },
+      },
+      {
+        id: 'data-filter',
+        name: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.DataFiltering', dm: '数据筛选' }),
+        required: true,
+        GI_CONTAINER: {
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.IntegratedComponents', dm: '集成组件' }),
+          type: 'string',
+          'x-decorator': 'FormItem',
+          'x-component': 'Select',
+          'x-component-props': {
+            mode: 'multiple',
+          },
+          enum: GIAC_CONTENT_ITEMS,
+          default: [],
+        },
+        icon: {
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.DisplayIcon', dm: '展示图标' }),
+          type: 'string',
+          'x-decorator': 'FormItem',
+          'x-component': 'Input',
+          default: 'icon-filter',
+        },
       },
       {
         id: 'styling-setting',
-        name: '布局样式',
+        name: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.LayoutStyle', dm: '布局样式' }),
         required: true,
         GI_CONTAINER: {
-          title: '集成组件',
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.IntegratedComponents', dm: '集成组件' }),
           type: 'string',
           'x-decorator': 'FormItem',
           'x-component': 'Select',
@@ -84,10 +122,10 @@ const registerMeta = context => {
       },
       {
         id: 'canvas-operator',
-        name: '画布操作',
+        name: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.CanvasOperation', dm: '画布操作' }),
         required: true,
         GI_CONTAINER: {
-          title: '集成组件',
+          title: $i18n.get({ id: 'basic.components.RichContainer.registerMeta.IntegratedComponents', dm: '集成组件' }),
           type: 'string',
           'x-decorator': 'FormItem',
           'x-component': 'Select',
@@ -99,6 +137,13 @@ const registerMeta = context => {
         },
       },
     ],
+    isSheet: {
+      title: '多页签',
+      type: 'string',
+      'x-decorator': 'FormItem',
+      'x-component': 'Switch',
+      default: false,
+    },
   };
 };
 export default registerMeta;