+++ packages/gi-common-components/src/NodeSelectionWrap/index.tsx
@@ -3,12 +3,13 @@
  * 用于获取目标节点，提供 2 种方式：从列表中选择节点或从画布中点选节点
  */
 import { FormOutlined } from '@ant-design/icons';
-import { Form, Select } from 'antd';
-import type { FormInstance } from 'antd';
-import React, { memo, useMemo, useState } from 'react';
-import './index.less';
 import { Graph } from '@antv/g6';
+import type { FormInstance } from 'antd';
+import { Button, Form, Select, Tooltip } from 'antd';
+import React, { memo, useMemo, useState } from 'react';
+import { Icon } from '../Icon';
 import $i18n from '../i18n';
+import './index.less';
 
 export enum NodeSelectionMode {
   List = 'List', // 从列表中选择节点
@@ -31,6 +32,7 @@ interface NodeSelectionWrapProps extends NodeSelectionProps {
 }
 
 interface NodeSelectionFormItemProps extends NodeSelectionProps {
+  color: string;
   key: string;
   name: string;
   label: string;
@@ -41,13 +43,13 @@ interface NodeSelectionFormItemProps extends NodeSelectionProps {
 let nodeClickListener = e => {};
 
 const NodeSelectionFormItem: React.FC<NodeSelectionFormItemProps> = memo(props => {
-  const { graph, nodeSelectionMode, nodeLabel, key, name, label, form, data, setSelecting, selecting } = props;
+  const { graph, nodeSelectionMode, nodeLabel, key, name, label, form, data, setSelecting, selecting, color } = props;
   const isList = nodeSelectionMode.includes(NodeSelectionMode.List);
   const isCanvas = nodeSelectionMode.includes(NodeSelectionMode.Canvas);
 
   const beginSelect = () => {
     setSelecting(name);
-    graph.off('node:click', nodeClickListener);
+    graph && graph.off('node:click', nodeClickListener);
 
     nodeClickListener = e => {
       setSelecting('');
@@ -75,10 +77,11 @@ const NodeSelectionFormItem: React.FC<NodeSelectionFormItemProps> = memo(props =
         className="main"
         key={key}
         name={name}
-        label={label}
+        label={<Icon type="icon-dot" style={{ fontSize: '12px', color }} />}
+        colon={false}
         rules={[
           {
-            required: true,
+            // required: true,
             message: $i18n.get(
               {
                 id: 'common-components.src.NodeSelectionWrap.PleaseFillInLabelNodelabel',
@@ -88,18 +91,6 @@ const NodeSelectionFormItem: React.FC<NodeSelectionFormItemProps> = memo(props =
             ),
           },
         ]}
-        tooltip={
-          isCanvas && {
-            open: selecting === name,
-            title: $i18n.get(
-              {
-                id: 'common-components.src.NodeSelectionWrap.YouCanClickCanvasNodes',
-                dm: `可点选画布节点，快速选择${label}`,
-              },
-              { label: label },
-            ),
-          }
-        }
       >
         <Select
           showSearch
@@ -117,11 +108,21 @@ const NodeSelectionFormItem: React.FC<NodeSelectionFormItemProps> = memo(props =
         </Select>
       </Form.Item>
       {isCanvas && (
-        <FormOutlined
-          className="operation"
-          style={{ cursor: 'pointer', color: selecting === name ? '#1890ff' : 'rgba(0, 0, 0, 0.65)' }}
-          onClick={beginSelect}
-        />
+        <Tooltip
+          title={$i18n.get(
+            {
+              id: 'common-components.src.NodeSelectionWrap.YouCanClickCanvasNodes',
+              dm: `可点选画布节点，快速选择${label}`,
+            },
+            { label: label },
+          )}
+        >
+          <FormOutlined
+            className="operation"
+            style={{ cursor: 'pointer', color: selecting === name ? '#1890ff' : 'rgba(0, 0, 0, 0.65)' }}
+            onClick={beginSelect}
+          />
+        </Tooltip>
       )}
     </div>
   );
@@ -130,10 +131,19 @@ const NodeSelectionFormItem: React.FC<NodeSelectionFormItemProps> = memo(props =
 const NodeSelectionWrap: React.FC<NodeSelectionWrapProps> = memo(props => {
   const { graph, nodeSelectionMode, nodeLabel, items, form, data } = props;
   const [selecting, setSelecting] = useState('');
+  const colors = ['#1650FF', '#FFC53D'];
+  const handleSwap = async () => {
+    const values = await form.getFieldsValue();
+    console.log('values', values);
+    const { source, target } = values;
+    form.setFieldsValue({ source: target, target: source });
+  };
+
   return (
-    <>
-      {items.map(item => (
+    <div style={{ position: 'relative' }}>
+      {items.map((item, index) => (
         <NodeSelectionFormItem
+          color={colors[index]}
           graph={graph}
           form={form}
           key={item.name}
@@ -146,7 +156,35 @@ const NodeSelectionWrap: React.FC<NodeSelectionWrapProps> = memo(props => {
           setSelecting={setSelecting}
         />
       ))}
-    </>
+      <div
+        className="gi-path-analysis-line"
+        style={{
+          position: 'absolute',
+          top: '22px',
+          left: '-6px',
+          height: '44px',
+          justifyContent: 'center',
+          alignItems: 'center',
+          display: 'flex',
+        }}
+      >
+        <div
+          style={{
+            height: '38px',
+            width: '1px',
+            background: '#DDDDDF',
+            position: 'absolute',
+          }}
+        ></div>
+        <Button
+          icon={<Icon type="icon-swap" />}
+          size="small"
+          type="text"
+          onClick={handleSwap}
+          style={{ background: '#fff' }}
+        />
+      </div>
+    </div>
   );
 });